+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ##### Unreleased
 
 * Changes `whitelist`/`blacklist` to `allowlist`/`blocklist` to follow [guidelines](https://chromium.googlesource.com/chromium/src/+/master/styleguide/inclusive_code.md#racially-neutral)
+* Updates `allowlist`, `blocklist` csrf config to allow specifying the type of match required
+
 
 
 ##### v1.6.1